Use array instead of list, unwrap process loop
Since we know there's exactly two items in the collection, we should
just use a simple array. Also, since there's always exactly two entries
in the array, there's no need to foreach the array, but rather we can
just unroll the loop preventing n calls to GetEnumerator and all of
that infrastructure.

This provided a small performance benefit.