Add example of using EWrapper implementation
This is meant as example code. It was NOT run. I've provided an
example implementation of the GetExecutions method. If you look
at the diff against origin/master for this function you'll see
minimal changes were actually made (mostly just types). This is
by design. Ideally the diff will look very small since we've
added the EWrapper implementation as a shim against the old
event based implementation.

If you have any questions, please don't hesitate to reach out
for guidance.